+++ lib/nib/code_climate.rb
@@ -12,24 +12,15 @@ class Nib::CodeClimate
 
   def script
     @script ||= <<-SCRIPT
-      docker-compose \
-        -f #{compose_file.path} \
-        run \
+      docker run \
+        --interactive \
+        --tty \
         --rm \
-        codeclimate \
-        #{command}
+        --env CODECLIMATE_CODE="$PWD" \
+        --volume "$PWD":/code \
+        --volume /var/run/docker.sock:/var/run/docker.sock \
+        --volume /tmp/cc:/tmp/cc \
+        codeclimate/codeclimate #{command || 'help'}
     SCRIPT
-  end
-
-  private
-
-  def config
-    @config ||= Nib.load_config(:codeclimate, 'docker-compose.yml')
-  end
-
-  def compose_file
-    @compose_file ||= Tempfile.open('compose') do |file|
-      file.tap { |f| f.write(config) }
-    end
   end
 end